chore: Async Storage fetching

+++ src/storage/Days/dayRegister.ts
@@ -4,14 +4,21 @@ import AsyncStorage from "@react-native-async-storage/async-storage";
 import { DAYS_COLLECTION } from "@storage/storageConfig";
 import { daysGetAll } from "./daysGetAll";
 
-export async function dayRegister(newDate: string){
-  try{
+export async function dayRegister(newDate: string) {
+  try {
 
     const storedDays = await daysGetAll();
 
-    const storage = JSON.stringify([...storedDays ,newDate]);
-    await AsyncStorage.setItem(DAYS_COLLECTION, storage);
-  }catch(error){
+    const dayAlreadyRegistered = storedDays.filter(day => day === newDate);
 
+    if (dayAlreadyRegistered.length > 0) {
+
+    } else {
+      const storage = JSON.stringify([...storedDays, newDate]);
+      await AsyncStorage.setItem(DAYS_COLLECTION, storage);
+    }
+
+  } catch (error) {
+    throw (error)
   }
 }
